% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{check_fields}
\alias{check_fields}
\title{Internal check for fields}
\usage{
check_fields(x, what = c("source", "id", "taxonomy", "sequence", "marker"))
}
\arguments{
\item{x}{a reference database (tibble object).}

\item{what}{a vector of fields to be checked.}
}
\value{
Invisible or error.
}
\description{
Internal check for fields
}
