% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_NCBI.R
\name{get_ncbi_taxonomy}
\alias{get_ncbi_taxonomy}
\title{Get NCBI taxonomy}
\usage{
get_ncbi_taxonomy(id)
}
\arguments{
\item{id}{A vector of id for records in the NCBI Taxonomy database.}
}
\value{
A tibble with each row corresponding to an id and each column
to a taxonomic level.
}
\description{
Download and parse NCBI taxonomy records
}
