% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_plots.R
\name{igraph_from_taxo}
\alias{igraph_from_taxo}
\title{Create a graph from a taxonomic table}
\usage{
igraph_from_taxo(x, cols = NULL)
}
\arguments{
\item{x}{a reference database (tibble).}

\item{cols}{an optional vector of column names to use a subset of columns.}
}
\value{
An \pkg{igraph} object representing taxonomic relationships.
}
\description{
Create a graph representation from a taxonomic classification
included in a reference
database. For this function to work, taxonomic fields must be set.
}
