% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_NCBI.R
\name{make_ncbi_table}
\alias{make_ncbi_table}
\title{Parse NCBI XML and make a table}
\usage{
make_ncbi_table(x)
}
\arguments{
\item{x}{A XML nodeset.}
}
\value{
A tibble.
}
\description{
Parse NCBI XML and make a table
}
