% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_ncbi_taxonomy.R
\name{ncbi_taxo_rank}
\alias{ncbi_taxo_rank}
\title{Taxonomic ranks of the NCBI Taxonomy database}
\usage{
ncbi_taxo_rank()
}
\value{
a vector of ordered ranks
}
\description{
Taxonomic ranks of the NCBI Taxonomy database
}
