% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_NCBI.R
\name{process_geo_ncbi}
\alias{process_geo_ncbi}
\title{Process coordinate column returned by NCBI}
\usage{
process_geo_ncbi(x, col = "lat_lon")
}
\arguments{
\item{x}{NCBI dataframe.}

\item{col}{column name containing geographical coordinates.}
}
\value{
NCBI dataframe.
}
\description{
Process coordinate column returned by NCBI
}
