% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb-package.R
\docType{package}
\name{refdb-package}
\alias{refdb}
\alias{refdb-package}
\title{refdb: What the Package Does (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fkeck.github.io/refdb/}
  \item Report bugs at \url{https://github.com/fkeck/refdb/issues}
}

}
\author{
\strong{Maintainer}: Francois Keck \email{francois.keck@gmail.com} (\href{https://orcid.org/0000-0002-3323-4167}{ORCID}) [copyright holder]

}
\keyword{internal}
