% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_clean.R
\name{refdb_clean_seq_remove_gaps}
\alias{refdb_clean_seq_remove_gaps}
\title{Remove gaps from genetic sequences}
\usage{
refdb_clean_seq_remove_gaps(x)
}
\arguments{
\item{x}{a reference database with a defined sequence field.}
}
\value{
A reference database.
}
\description{
Remove gaps from genetic sequences
}
