% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_clean.R
\name{refdb_clean_seq_remove_sideN}
\alias{refdb_clean_seq_remove_sideN}
\title{Remove repeated side N from genetic sequences}
\usage{
refdb_clean_seq_remove_sideN(x, side = "both")
}
\arguments{
\item{x}{a reference database with a defined sequence field.}

\item{side}{which side to clean.
Can be one of \code{"left"}, \code{"right"} or \code{"both"} (default).}
}
\value{
A reference database.
}
\description{
Remove repeated side N from genetic sequences
}
\examples{
lib <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib <- refdb_set_fields_BOLD(lib)
refdb_clean_seq_remove_sideN(lib)

}
