% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_clean.R
\name{refdb_clean_tax_harmonize_nomenclature}
\alias{refdb_clean_tax_harmonize_nomenclature}
\title{Harmonize taxonomic name nomenclature}
\usage{
refdb_clean_tax_harmonize_nomenclature(x, cols = NULL)
}
\arguments{
\item{x}{a reference database.}

\item{cols}{an optional vector of column names.
If \code{NULL} (default), the function is applied to the columns
associated with the \code{taxonomy} and \code{organism} fields.}
}
\value{
A reference database.
}
\description{
Harmonize taxonomic name nomenclature
}
\examples{
lib <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib <- refdb_set_fields_BOLD(lib)
refdb_clean_tax_harmonize_nomenclature(lib)

}
