% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_clean.R
\name{refdb_clean_tax_remove_blank}
\alias{refdb_clean_tax_remove_blank}
\title{Remove blank characters from taxonomic names}
\usage{
refdb_clean_tax_remove_blank(x, cols = NULL)
}
\arguments{
\item{x}{a reference database.}

\item{cols}{an optional vector of column names.
If \code{NULL} (default), the function is applied to the columns
associated with the \code{taxonomy} and \code{organism} fields.}
}
\value{
A reference database.
}
\description{
Remove blank characters from taxonomic names
}
