% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_export.R
\name{refdb_export_utax}
\alias{refdb_export_utax}
\title{Export reference database for USEARCH/VSEARCH}
\usage{
refdb_export_utax(x, file, verbose = TRUE)
}
\arguments{
\item{x}{a reference database.}

\item{file}{a file path. This will be used to create a .fasta file.}

\item{verbose}{print information in the console.}
}
\value{
No return value, called for side effects.
}
\description{
Write a reference database in utax format.
}
\examples{
lib <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib <- refdb_set_fields_BOLD(lib)
refdb_export_utax(lib, tempfile())

}
