% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_fill_taxo.R
\name{refdb_fill_tax_downstream}
\alias{refdb_fill_tax_downstream}
\title{Fill missing data in taxonomy}
\usage{
refdb_fill_tax_downstream(x, qualifier = "indet.")
}
\arguments{
\item{x}{a reference database.}

\item{qualifier}{a string to add the new labels.
Default ensure that \code{refdb_clean_tax_NA} will correctly
identify the label as NA.}
}
\description{
Replace NA values in taxonomic classification
using upstream ranks.
}
\seealso{
refdb_fill_tax_upstream to replace NA values
using downstream data.
}
