% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_filter.R
\name{refdb_filter_seq_ambiguous}
\alias{refdb_filter_seq_ambiguous}
\title{Filter sequences based on their number of ambiguous character.}
\usage{
refdb_filter_seq_ambiguous(x, max_ambig = 3L, char = "N")
}
\arguments{
\item{x}{a reference database.}

\item{max_ambig}{maximum number of ambiguous character.}

\item{char}{characters interpreted as ambiguous (vector).}
}
\value{
A tibble (filtered reference database).
}
\description{
Filter sequences based on their number of ambiguous character.
}
