% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_filter.R
\name{refdb_filter_tax_na}
\alias{refdb_filter_tax_na}
\title{Filter records NA taxa}
\usage{
refdb_filter_tax_na(x)
}
\arguments{
\item{x}{a reference database.}

\item{min_tax}{minimum taxonomic level
(column name of the reference database).}
}
\value{
A tibble (filtered reference database).
}
\description{
Remove records where taxa is NA if it is not
the only representant of the upper clade.
Note that the function maybe slow on large datasets.
//EXPERIMENTAL//
}
