% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{refdb_get_fields}
\alias{refdb_get_fields}
\title{Get fields of a reference database}
\usage{
refdb_get_fields(x, silent = FALSE)
}
\arguments{
\item{x}{a reference database.}

\item{silent}{if \code{TRUE} silently and invisibly
returns fields.}
}
\value{
The list of fields is returned invisibly.
}
\description{
Get fields of a reference database
}
