% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_sample.R
\name{refdb_sample_tax}
\alias{refdb_sample_tax}
\title{Sample records within taxa}
\usage{
refdb_sample_tax(x, n_max = 10, cols = NULL)
}
\arguments{
\item{x}{a reference database.}

\item{n_max}{maximum number of records to keep for each taxa.}

\item{cols}{an optional vector of column names.
If \code{NULL} (default), the function is applied to the columns
associated with the \code{taxonomy} field.}
}
\value{
A reference database.
}
\description{
This function can be useful to keep a maximum of records per taxa.
This function require dplyr dev version to work because of slice_sample.
Will be exported once available.
}
