% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{refdb_write_fields}
\alias{refdb_write_fields}
\title{Write fields to a file}
\usage{
refdb_write_fields(x, file)
}
\arguments{
\item{x}{a reference database with some fields to be saved.}

\item{file}{a path to the file to write.}
}
\value{
No return value, called for its side effects.
}
\description{
This function can be used to save fields defined
using e.g. \code{refdb_set_fields} to a file.
Data are saved in YAML and can be read again using the
\code{config_yaml} argument of \code{refdb_set_fields}.
}
\examples{
lib <- read.csv(system.file("extdata", "ephem.csv", package = "refdb"))
tmp <- tempfile()
refdb_write_fields(lib, tmp)

}
