% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_ncbi_taxonomy.R
\name{valid_taxo_rank}
\alias{valid_taxo_rank}
\title{Ranks considered as valid by refdb}
\usage{
valid_taxo_rank()
}
\value{
a vector of ordered ranks.
}
\description{
Ranks considered as valid by refdb
}
\examples{
valid_taxo_rank()

}
\references{
This is a simplified version of the
list \code{rank_ref} available in \pkg{taxize}.
}
