% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_NCBI.R
\name{xml_extract}
\alias{xml_extract}
\title{Extract XML elements}
\usage{
xml_extract(x, xpath)
}
\arguments{
\item{x}{A document, node, or node set.}

\item{xpath}{A string containing a xpath expression.}
}
\value{
A character vector, the same length as x.
}
\description{
Combine xml_find_first and xml_text to extract elements.
}
