% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_filter.R
\name{filter_scores}
\alias{filter_scores}
\alias{.filter_seq_length}
\title{Scores for filtering operations}
\usage{
.filter_seq_length(x, gaps)
}
\arguments{
\item{x}{a reference database}

\item{gaps}{should gaps be included.}
}
\value{
a numeric vector
}
\description{
Scores for filtering operations
}
