% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb-package.R
\docType{package}
\name{refdb-package}
\alias{refdb}
\alias{refdb-package}
\title{refdb: A DNA Reference Library Manager}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Reference database manager offering a set of functions to import, organize, clean, filter, audit and export reference genetic data. Provide functions to download sequence data from Bold Systems (\url{https://www.boldsystems.org/}) and NCBI GenBank \url{https://www.ncbi.nlm.nih.gov/genbank/}. Designed as an environment for semi-automatic and assisted construction of reference databases and to improve standardization and repeatability in barcoding and metabarcoding studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fkeck.github.io/refdb/}
  \item Report bugs at \url{https://github.com/fkeck/refdb/issues}
}

}
\author{
\strong{Maintainer}: Francois Keck \email{francois.keck@gmail.com} (\href{https://orcid.org/0000-0002-3323-4167}{ORCID}) [copyright holder]

}
\keyword{internal}
