% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_checks.R
\name{refdb_check_seq_conflict}
\alias{refdb_check_seq_conflict}
\title{Check for conflicts in sequences}
\usage{
refdb_check_seq_conflict(x)
}
\arguments{
\item{x}{a reference database.}
}
\value{
A list of two-columns tibbles reporting duplicated
sequences with different taxonomy.
}
\description{
Check for conflicts in sequences
}
