% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_clean.R
\name{refdb_clean_seq_crop_primers}
\alias{refdb_clean_seq_crop_primers}
\title{Crop genetic sequences with a set of primers}
\usage{
refdb_clean_seq_crop_primers(
  x,
  pattern_in,
  pattern_out,
  max_error_in = 0.1,
  max_error_out = 0.1,
  include_primers = TRUE
)
}
\arguments{
\item{x}{a reference database with a defined sequence field.}

\item{pattern_out}{}
}
\value{
A reference database.
}
\description{
Crop genetic sequences with a set of primers
}
