% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_export.R
\name{refdb_export_mothur}
\alias{refdb_export_mothur}
\title{Export reference database for Mothur}
\usage{
refdb_export_mothur(x, file)
}
\arguments{
\item{x}{a reference database.}

\item{file}{a file path. This will be used to create
a .fasta file and a .txt file.}
}
\value{
No return value, called for side effects.
}
\description{
Write a reference database in formats which can be used
with \code{Mothur}.
}
\examples{
lib <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib <- refdb_set_fields_BOLD(lib)
refdb_export_mothur(lib, tempfile())

}
