% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_filter.R
\name{refdb_filter_seq_duplicates}
\alias{refdb_filter_seq_duplicates}
\title{Filter duplicated sequences.}
\usage{
refdb_filter_seq_duplicates(x)
}
\arguments{
\item{x}{a reference database.}
}
\value{
A tibble (filtered reference database).
}
\description{
Exclude duplicated sequences.
This is based both on sequences and taxonomy.
NA values are assumed to be comparable.
}
