% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_filter.R
\name{refdb_filter_seq_homopolymers}
\alias{refdb_filter_seq_homopolymers}
\title{Filter sequences based on their number of repeated character.}
\usage{
refdb_filter_seq_homopolymers(x, max_len = 16L)
}
\arguments{
\item{x}{a reference database.}

\item{max_len}{maximum number of repeated character (homopolymer).}
}
\value{
A tibble (filtered reference database).
}
\description{
Filter sequences based on their number of repeated character.
}
\examples{
lib <- read.csv(system.file("extdata", "baetidae_bold.csv", package = "refdb"))
lib <- refdb_set_fields_BOLD(lib)
refdb_filter_seq_homopolymers(lib)

}
