% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdb_filter.R
\name{refdb_filter_seq_length}
\alias{refdb_filter_seq_length}
\title{Filter sequences based on their number of character.}
\usage{
refdb_filter_seq_length(x, min_len, max_len)
}
\arguments{
\item{x}{a reference database.}

\item{min_len}{minimum sequence length.}

\item{max_len}{maximum sequence length.}
}
\value{
A tibble (filtered reference database).
}
\description{
Filter sequences based on their number of character.
}
